Move CartPole renderer under continuous envs
