+++ masa/envs/continuous/renderers/__init__.py
@@ -1,0 +1,2 @@
+"""Renderers for continuous-action environments."""
+
